final commit for targil 5

+++ include/Coordinates.h
@@ -2,13 +2,13 @@
 #include <string>
 class Coordinates
 {
-	int y_cord;
-	int x_cord;
+	float y_cord;
+	float x_cord;
 	public:
 		Coordinates() : y_cord(0), x_cord(0) {}
 
-		int getX() const { return x_cord; }
-		int getY() const { return y_cord; }
-		void setX(int x) { x_cord = x; }
-		void setY(int y) { y_cord = y; }
+		float getX() const { return x_cord; }
+		float getY() const { return y_cord; }
+		void setX(float x) { x_cord = x; }
+		void setY(float y) { y_cord = y; }
 };
